:hammer: refactor root dir name #4
