+++ src/main/java/Backend/teampple/domain/domain.java
@@ -1,0 +1,4 @@
+package Backend.teampple.domain;
+
+public class domain {
+}
