+++ src/main/java/Backend/teampple/global/common/entity/DeleteStatus.java
@@ -1,0 +1,5 @@
+package Backend.teampple.global.common.entity;
+
+public enum DeleteStatus {
+    YES,NO
+}
