+++ src/main/java/Backend/teampple/global/config/SwaggerConfig.java
@@ -1,0 +1,76 @@
+package Backend.teampple.global.config;
+
+import org.springframework.context.annotation.Bean;
+import org.springframework.context.annotation.Configuration;
+import org.springframework.context.annotation.Profile;
+import org.springframework.web.servlet.config.annotation.EnableWebMvc;
+import springfox.documentation.builders.ApiInfoBuilder;
+import springfox.documentation.builders.PathSelectors;
+import springfox.documentation.builders.RequestHandlerSelectors;
+import springfox.documentation.service.*;
+import springfox.documentation.spi.DocumentationType;
+import springfox.documentation.spi.service.contexts.SecurityContext;
+import springfox.documentation.spring.web.plugins.Docket;
+
+import java.util.*;
+
+@Configuration
+@EnableWebMvc
+@Profile(value = "!prod")
+public class SwaggerConfig {
+
+    private ApiInfo swaggerInfo() {
+        return new ApiInfoBuilder()
+                .title("Teampple Api Docs")
+                .version("0.0.1")
+                .description("Teampple Api 문서입니다")
+                .build();
+    }
+
+    @Bean
+    public Docket swaggerApi() {
+        return new Docket(DocumentationType.SWAGGER_2)
+                .securityContexts(Arrays.asList(securityContext()))
+                .securitySchemes(Arrays.asList(apiKey()))
+                .consumes(getConsumeContentTypes())
+                .produces(getProduceContentTypes())
+                .apiInfo(swaggerInfo())
+                .select()
+                .apis(RequestHandlerSelectors.basePackage("Backend.teampple")) /**하위 패키지 -> API*/
+                .paths(PathSelectors.ant("/api/**")) /** "/api/**"인 url 들만 필터링*/
+                .build();
+    }
+
+    private SecurityContext securityContext() {
+        return SecurityContext.builder()
+                .securityReferences(defaultAuth())
+                .build();
+    }
+    private List<SecurityReference> defaultAuth() {
+        AuthorizationScope authorizationScope = new AuthorizationScope(
+                "global",
+                "accessEverything"
+        );
+        AuthorizationScope[] authorizationScopes = new AuthorizationScope[1];
+        authorizationScopes[0] = authorizationScope;
+        return Arrays.asList(new SecurityReference("Authorization", authorizationScopes));
+    }
+    private ApiKey apiKey() {
+        return new ApiKey("Authorization", "X-AUTH-TOKEN", "header");
+    }
+    private Set<String> getConsumeContentTypes() {
+        Set<String> consumes = new HashSet<>();
+        consumes.add("application/json;charset=UTF-8");
+        consumes.add("application/x-www-form-urlencoded");
+        return consumes;
+    }
+
+    private Set<String> getProduceContentTypes() {
+        Set<String> produces = new HashSet<>();
+        produces.add("application/json;charset=UTF-8");
+        return produces;
+    }
+
+}
+//swagger: http://localhost:8080/api/swagger-ui/index.html
+//swagger: https://www.teampple.site/api/swagger-ui/index.html
