+++ src/main/java/Backend/teampple/global/error/ErrorCode.java
@@ -1,0 +1,26 @@
+package Backend.teampple.global.error;
+
+import lombok.AllArgsConstructor;
+import lombok.Getter;
+import org.springframework.http.HttpStatus;
+
+import static org.springframework.http.HttpStatus.*;
+
+@Getter
+@AllArgsConstructor
+public enum ErrorCode {
+
+    // common
+    INVALID_INPUT_VALUE(BAD_REQUEST, "C001", " Invalid Input Value"),
+    _INTERNAL_SERVER_ERROR(INTERNAL_SERVER_ERROR, "C000", "서버 에러, 관리자에게 문의 바랍니다."),
+    _BAD_REQUEST(BAD_REQUEST, "C001", "잘못된 요청입니다."),
+    _UNAUTHORIZED(UNAUTHORIZED, "C002", "권한이 없습니다."),
+    _METHOD_NOT_ALLOWED(METHOD_NOT_ALLOWED, "C003", "지원하지 않는 Http Method 입니다."),
+
+    // test
+    TEST(BAD_REQUEST, "T001", "test");
+
+    private final HttpStatus httpStatus;
+    private final String code;
+    private final String message;
+}
