+++ src/main/java/Backend/teampple/global/error/exception/BadRequestException.java
@@ -1,0 +1,18 @@
+package Backend.teampple.global.error.exception;
+
+
+import Backend.teampple.global.error.ErrorCode;
+import lombok.Getter;
+
+/**
+ * status : 400
+ */
+@Getter
+public class BadRequestException extends BaseException {
+    public BadRequestException() {
+        super(ErrorCode._BAD_REQUEST, ErrorCode._BAD_REQUEST.getMessage());
+    }
+    public BadRequestException(String message) {
+        super(ErrorCode._BAD_REQUEST, message);
+    }
+}
