--- src/main/java/Backend/teampplus/TeamppleApplication.java
@@ -1,13 +1,0 @@
-package Backend.teampplus;
-
-import org.springframework.boot.SpringApplication;
-import org.springframework.boot.autoconfigure.SpringBootApplication;
-
-@SpringBootApplication
-public class TeamppleApplication {
-
-	public static void main(String[] args) {
-		SpringApplication.run(TeamppleApplication.class, args);
-	}
-
-}
