--- src/main/java/Backend/teampplus/domain/domain.java
@@ -1,4 +1,0 @@
-package Backend.teampplus.domain;
-
-public class domain {
-}
