--- src/main/java/Backend/teampplus/global/common/entity/DeleteStatus.java
@@ -1,5 +1,0 @@
-package Backend.teampplus.global.common.entity;
-
-public enum DeleteStatus {
-    YES,NO
-}
