--- src/main/java/Backend/teampplus/global/config/ModelMapperConfig.java
@@ -1,13 +1,0 @@
-package Backend.teampplus.global.config;
-
-import org.modelmapper.ModelMapper;
-import org.springframework.context.annotation.Bean;
-import org.springframework.context.annotation.Configuration;
-
-@Configuration
-public class ModelMapperConfig {
-    @Bean
-    public ModelMapper modelMapper(){
-        return new ModelMapper();
-    }
-}
