--- src/main/java/Backend/teampplus/global/config/S3Config.java
@@ -1,51 +1,0 @@
-package Backend.teampplus.global.config;
-
-import com.amazonaws.auth.AWSCredentials;
-import com.amazonaws.auth.AWSStaticCredentialsProvider;
-import com.amazonaws.auth.BasicAWSCredentials;
-import com.amazonaws.services.s3.AmazonS3;
-import com.amazonaws.services.s3.AmazonS3ClientBuilder;
-import com.amazonaws.services.s3.transfer.TransferManager;
-import com.amazonaws.services.s3.transfer.TransferManagerBuilder;
-import org.springframework.beans.factory.annotation.Value;
-import org.springframework.context.annotation.Bean;
-import org.springframework.context.annotation.Configuration;
-import org.springframework.context.annotation.Primary;
-
-@Configuration
-public class S3Config {
-
-    @Value("${cloud.aws.credentials.access-key}")
-    private String accessKey;
-
-    @Value("${cloud.aws.credentials.secret-key}")
-    private String secretKey;
-
-    @Value("${cloud.aws.region.static}")
-    private String region;
-
-    @Bean
-    @Primary
-    public BasicAWSCredentials awsCredentialsProvider(){
-        return new BasicAWSCredentials(accessKey, secretKey);
-    }
-
-    @Bean
-    public AmazonS3 amazonS3Client() {
-        AWSCredentials credentials = awsCredentialsProvider();
-
-        return AmazonS3ClientBuilder
-                .standard()
-                .withCredentials(new AWSStaticCredentialsProvider(credentials))
-                .withRegion(region)
-                .build();
-    }
-
-    @Bean
-    public TransferManager transferManager() {
-        return TransferManagerBuilder.standard()
-                .withS3Client(amazonS3Client())
-                .build();
-    }
-
-}
