+++ src/test/java/Backend/teampple/TeamppleApplicationTests.java
@@ -1,0 +1,13 @@
+package Backend.teampple;
+
+import org.junit.jupiter.api.Test;
+import org.springframework.boot.test.context.SpringBootTest;
+
+@SpringBootTest
+class TeamppleApplicationTests {
+
+	@Test
+	void contextLoads() {
+	}
+
+}
