--- src/test/java/Backend/teampplus/TeamppleApplicationTests.java
@@ -1,13 +1,0 @@
-package Backend.teampplus;
-
-import org.junit.jupiter.api.Test;
-import org.springframework.boot.test.context.SpringBootTest;
-
-@SpringBootTest
-class TeamppleApplicationTests {
-
-	@Test
-	void contextLoads() {
-	}
-
-}
